# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Insurance Policy Simplifier API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Policy {
  id: number;
  filename: string;
  policy_name: string;
  policy_type: string;
  simple_explanation: string;
  claim_difficulty_score: string;
  coverage: string[];
  exclusions: string[];
  claim_process: string[];
  important_warnings: string[];
  waiting_periods: string[];
  created_at: string;
}

export interface PolicyTypeCount {
  type: string;
  count: number;
}

export interface PolicyStats {
  total_policies: number;
  policy_types: PolicyTypeCount[];
  avg_difficulty_score: string;
  recent_policies: Policy[];
}

export type PolicyExplainRequestLanguage = typeof PolicyExplainRequestLanguage[keyof typeof PolicyExplainRequestLanguage];


export const PolicyExplainRequestLanguage = {
  en: 'en',
  hi: 'hi',
  mr: 'mr',
} as const;

export interface PolicyExplainRequest {
  language: PolicyExplainRequestLanguage;
}

export interface PolicyExplanation {
  text: string;
  language: string;
}