# lib/api-zod/src/generated/types/policyExplainRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Insurance Policy Simplifier API
 * OpenAPI spec version: 0.1.0
 */
import type { PolicyExplainRequestLanguage } from './policyExplainRequestLanguage';

export interface PolicyExplainRequest {
  language: PolicyExplainRequestLanguage;
}